adapt to 64-bit

+++ src/MAddResIDDlg.hpp
@@ -68,7 +68,8 @@ public:
         SetDlgItemTextW(hwnd, edt1, table[IDTYPE_default].name.c_str());
         m_bChanging = FALSE;
 
-        SendDlgItemMessage(hwnd, scr1, UDM_SETRANGE, 0, MAKELPARAM(SHRT_MAX, SHRT_MIN));
+        SendDlgItemMessage(hwnd, scr1, UDM_SETRANGE, 0,
+                           MAKELPARAM((WORD)SHRT_MAX, (WORD)SHRT_MIN));
 
         CenterWindowDx();
         return TRUE;
